requirements.txt
aiodns==3.0.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
ccxt==3.1.52
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
emoji==2.5.1
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
magic-filter==1.0.9
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
python-axolotl-curve25519==0.4.1.post2
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
zipp==3.15.0